refactor(parser): move Use AST node
